Extract shared smart-wallet-connect module + add CBSW ERC-1271 signing
Refactor: Coinbase Smart Wallet and DSProxy pages were ~95% duplicated
WalletConnect plumbing. Extracted the shared flow into
`app/wallet/_smart-wallet-connect/` parameterized by a `SmartWalletConfig`
(identity, owner check, tx wrapper, chain-switch behavior, optional sign
wrappers, footer). Each page is now ~8 LOC.

ERC-1271 signing: the original implementation had the EOA sign the dApp's
hash directly, which a smart wallet's `isValidSignature` would always
reject. Implement Coinbase Smart Wallet's `replaySafeHash` typed-data
path: sign EIP-712 over `CoinbaseSmartWalletMessage(bytes32 hash)` with
the wallet's domain, then ABI-encode as the `SignatureWrapper(uint256
ownerIndex, bytes signatureData)` that `isValidSignature` decodes. Look
up the EOA's owner index by scanning `ownerAtIndex` (CBSW stores EOAs as
`abi.encode(address)`). Modal compares signer to the smart-wallet
address (was incorrectly compared to the EOA). DSProxy intentionally
omits the sign callbacks since it's a forwarder, not an account; the
modal rejects with a clear error.
